Break down SQL tests in multiple files
