+++ sql/pom.xml
@@ -53,13 +53,6 @@
       <artifactId>junit</artifactId>
       <version>4.12</version>
     </dependency>
-    <!-- Uncomment if required
-    <dependency>
-      <groupId>ddlogapi</groupId>
-      <artifactId>ddlog</artifactId>
-      <version>1.0-SNAPSHOT</version>
-      <scope>test</scope>
-    </dependency> -->
   </dependencies>
   <properties>
     <maven.compiler.source>1.8</maven.compiler.source>
