+++ sql/src/main/java/com/vmware/ddlog/translator/TranslationException.java
@@ -16,6 +16,8 @@ import com.facebook.presto.sql.tree.Node;
 import javax.annotation.Nullable;
 
 public class TranslationException extends RuntimeException {
+    private static final long serialVersionUID = 1L;
+
     public TranslationException(String message, @Nullable Node node) {
         super(message + "\n" + getPosition(node));
     }
